The following script fails: kudu/build-support/mini-cluster/build_mini_cluster_binaries.sh
with the following error:
++ /kudu/kudu/build-support/mini-cluster/relocate_binaries_for_mini_cluster.py /kudu/kudu/build/mini-cluster kudu kudu-tserver kudu-master
2022-03-23 17:57:58,557 INFO: Including target 'kudu' out of '['kudu', 'kudu-tserver', 'kudu-master']' and its dependencies in archive...
2022-03-23 17:57:58,652 INFO: Attempting to relocate sasl2 modules...
Traceback (most recent call last):
  File "/kudu/kudu/build-support/mini-cluster/relocate_binaries_for_mini_cluster.py", line 501, in <module>
    main()
  File "/kudu/kudu/build-support/mini-cluster/relocate_binaries_for_mini_cluster.py", line 498, in main
    relocate_deps(target_src, target_dst, config)
  File "/kudu/kudu/build-support/mini-cluster/relocate_binaries_for_mini_cluster.py", line 404, in relocate_deps
    return relocate_deps_linux(target_src, target_dst, config)
  File "/kudu/kudu/build-support/mini-cluster/relocate_binaries_for_mini_cluster.py", line 340, in relocate_deps_linux
    libs = dep_extractor.extract_deps(target_src)
  File "/kudu/kudu/build-support/mini-cluster/../../build-support/dep_extract.py", line 108, in extract_deps
    match = LDD_RE.match(line)
TypeError: cannot use a string pattern on a bytes-like object

This error is traced to line 108 for the file
kudu/build-support/dep_extract.py

      match = LDD_RE.match(line)

The “re” function expects ASCII character string. In python 2.x the default treatment of charters follows ASCII standard. However, in python 3.x the default encoding follows the UTF-8 standard. This leads to a failure in python 3.X when the UTF-8 encoded character objects are used as arguments to functions such as “re”. The recommended fix is to decode the string before it is passed as an argument to the ‘re’ call, and this works with both python 2.x as well as python 3.x.

diff --git a/build-support/dep_extract.py b/build-support/dep_extract.py
index ba454b635..21c7b8d29 100644
--- a/build-support/dep_extract.py
+++ b/build-support/dep_extract.py
@@ -105,7 +105,7 @@ class DependencyExtractor(object):

     deps = []
     for line in out.splitlines():
-      match = LDD_RE.match(line)
+      match = LDD_RE.match(line.decode('utf-8'))
       if not match:
         continue
       dep = match.group(1)